restructure a part of tools/ folder
