+++ tests/feature/test_amino_acid.py
@@ -3,8 +3,8 @@ from pdb2sql import pdb2sql
 from deeprankcore.models.amino_acid import serine
 from deeprankcore.models.variant import SingleResidueVariant
 from deeprankcore.features.sasa import add_features
-from deeprankcore.tools.graph import build_residue_graph
-from deeprankcore.tools.pdb import get_structure, get_surrounding_residues
+from deeprankcore.models.graph import build_residue_graph
+from deeprankcore.models.pdb import get_structure, get_surrounding_residues
 from deeprankcore.domain.features import nodefeats
 from deeprankcore.features.amino_acid import add_features # noqa
