+++ tests/feature/test_atom.py
@@ -2,9 +2,9 @@ import numpy
 from pdb2sql import pdb2sql
 
 from deeprankcore.features.atom import add_features
-from deeprankcore.tools.graph import build_atomic_graph
+from deeprankcore.models.graph import build_atomic_graph
 from deeprankcore.models.structure import Chain, Residue
-from deeprankcore.tools.pdb import get_structure, get_surrounding_residues
+from deeprankcore.models.pdb import get_structure, get_surrounding_residues
 from deeprankcore.domain.features import nodefeats
