+++ tests/feature/test_atomic_contact.py
@@ -6,7 +6,7 @@ import numpy
 from deeprankcore.models.structure import Chain, Atom
 from deeprankcore.models.contact import AtomicContact, ResidueContact
 from deeprankcore.models.graph import Edge, Graph
-from deeprankcore.tools.pdb import get_structure
+from deeprankcore.models.pdb import get_structure
 from deeprankcore.features.atomic_contact import add_features
 from deeprankcore.models.amino_acid import alanine
 from deeprankcore.models.variant import SingleResidueVariant
