+++ tests/feature/test_bsa.py
@@ -1,8 +1,8 @@
 from pdb2sql import pdb2sql
 from deeprankcore.models.structure import Structure
 from deeprankcore.features.bsa import add_features
-from deeprankcore.tools.graph import build_residue_graph, build_atomic_graph
-from deeprankcore.tools.pdb import get_structure, get_residue_contact_pairs
+from deeprankcore.models.graph import build_residue_graph, build_atomic_graph
+from deeprankcore.models.pdb import get_structure, get_residue_contact_pairs
 from deeprankcore.domain.features import nodefeats
