--- tests/test_bsa.py
@@ -1,15 +1,0 @@
-import unittest
-from deeprankcore.tools.BSA import BSA
-
-
-class TestBSA(unittest.TestCase):
-    def setUp(self):
-        self.bsa = BSA("tests/data/pdb/1ATN/1ATN_1w.pdb")
-
-    def test_structure(self):
-        self.bsa.get_structure()
-        self.bsa.get_contact_residue_sasa()
-
-
-if __name__ == "__main__":
-    unittest.main()
