+++ tests/tools/test_pssm.py
@@ -1,6 +1,6 @@
 from pdb2sql import pdb2sql
 from deeprankcore.tools.pssm import parse_pssm
-from deeprankcore.tools.pdb import get_structure
+from deeprankcore.models.pdb import get_structure
 from deeprankcore.models.amino_acid import alanine
